move DateParser into separate file
