+++ android/app/build.gradle
@@ -41,7 +41,7 @@ android {
 
             testLogging {
                 events "skipped", "failed"
-                showStandardStreams = true
+                showStandardStreams = false // set to true to debug tests
                 exceptionFormat = "full"
             }
         }
@@ -72,4 +72,9 @@ dependencies {
 
     implementation "com.squareup.okhttp3:okhttp:5.3.2"
     implementation "io.github.g00fy2:versioncompare:1.5.0"
+
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.14.2'
+    testImplementation "io.mockk:mockk:1.14.9"
+    testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
+    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.14.2'
 }
